This release includes bug fixes,
including:

Fixed a problem where the app is not installed on the user device although the required activity runs directly on the emulator. It was being caused due to the missing code in the AndroidManifest.xml file, it required an intent-filter to indicate which activity was the main activity that would run on the user device by default.